chore: clean code by norminette
