+++ Makefile
@@ -20,8 +20,8 @@ COLOR				:=	get_color	checkers
 TEXTURE				:=	get_texture	cardinal_points
 VALIDATE_MAP		:=	validate_map	validate_player	validate_border	validate_char	\
 						validate_up_down	validate_left_right	validate_middle
-ENGINE				:=	draw	hooks	render_map					\
-						bresenham	render
+ENGINE				:=	draw	draw_2	calculations	calculations_2	hooks	\
+						render_map	bresenham	render
 
 SRC		:=	$(MAIN)	\
 			$(addprefix interpretate_map/, $(INTERPRETATE_MAP))	\
