+++ include/cub3d.h
@@ -69,12 +69,26 @@ void	walk_right(t_player *player);
 void	walk_forward(t_game *game);
 void	walk_left(t_player *player);
 
-//Draw
+// Draw
 void	draw_square(t_img *img, int x0, int y0, int color);
 void	draw_background(t_game game);
 void	pixel_put(t_img *data, int x, int y, int color);
 void	bresenham_x(t_img *img, int i[2], int f[2], int color);
 void	bresenham_y(t_img *img, int i[2], int f[2], int color);
+void	draw_x_line(t_game *game, t_ray_info *ray, int line_height, int x);
+void	draw_background(t_game game);
+
+// Calculations
+t_double_vector	calc_ray_dir(int x, t_game *game);
+void	calc_ray_info(t_game *game, t_ray_info *ray);
+void	calc_wall_hit(t_game *game, t_ray_info *ray);
+void	calc_perp_wall_dist(t_game *game, t_ray_info *ray);
+int	calc_lowest_pixel(int line_height);
+int	calc_highest_pixel(int line_height);
+double	calc_tile_hit_x(t_game *game, t_ray_info *ray);
+int	calc_tex_hit_x(t_game *game, t_ray_info *ray, t_img tex);
+unsigned long	tex_color(t_game *game, t_img tex, int tex_pos, int tex_hit_x);
+t_img	def_tex(t_game *game, t_ray_info *ray);
 
 // Utils
 void	print_mtx(t_game *game);
